+++ sql/NexoraDB/Tables/dbo.AccessProfile.sql
@@ -1,5 +1,7 @@
 ﻿USE [nexora]
 GO
+ALTER TABLE [dbo].[AccessProfile] DROP CONSTRAINT [FK_AccessProfile_Organizations]
+GO
 DROP TABLE [dbo].[AccessProfile]
 GO
 SET ANSI_NULLS ON
@@ -10,6 +12,7 @@ CREATE TABLE [dbo].[AccessProfile](
 	[AccessID] [int] IDENTITY(1,1) NOT NULL,
 	[Name] [nvarchar](50) NOT NULL,
 	[Description] [nvarchar](200) NULL,
+	[OrganizationCode] [nvarchar](5) NULL,
 PRIMARY KEY CLUSTERED 
 (
 	[AccessID] ASC
@@ -19,4 +22,9 @@ PRIMARY KEY CLUSTERED
 	[Name] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 ) ON [PRIMARY]
+GO
+ALTER TABLE [dbo].[AccessProfile]  WITH CHECK ADD  CONSTRAINT [FK_AccessProfile_Organizations] FOREIGN KEY([OrganizationCode])
+REFERENCES [dbo].[Organizations] ([organizationcode])
+GO
+ALTER TABLE [dbo].[AccessProfile] CHECK CONSTRAINT [FK_AccessProfile_Organizations]
 GO